remove zustand, refactor accordingly
