+++ package.json
@@ -27,8 +27,7 @@
     "react-query": "^3.39.1",
     "superjson": "^1.9.1",
     "uuid": "^8.3.2",
-    "zod": "^3.17.3",
-    "zustand": "4.0.0-rc.1"
+    "zod": "^3.17.3"
   },
   "devDependencies": {
     "@testing-library/react": "^13.3.0",
