+++ src/styles/globals.css
@@ -92,3 +92,8 @@ table {
   border-collapse: collapse;
   border-spacing: 0;
 }
+
+html {
+  height: 100vh;
+  min-height: 100vh;
+}
